pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "AstroBrain"
version = "0.1.0"
description = "AI-powered mission design and orbital analysis assistant with multi-agent architecture"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Om Yeole" }
]
maintainers = [
    { name = "Om Yeole" }
]
requires-python = ">=3.12,<3.13"
keywords = ["aerospace", "orbital-mechanics", "ai-agents", "langgraph", "mission-design", "satellite", "space"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Astronomy",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core async and HTTP
    "aiohttp>=3.13.2,<4.0.0",
    "aiohttp-retry>=2.9.1,<3.0.0",
    "httpx>=0.28.0,<0.29.0",
    
    # Scientific computing
    "numpy>=1.26.4,<2.0.0",
    "scipy>=1.16.0,<2.0.0",
    "pandas>=2.3.0,<3.0.0",
    "matplotlib>=3.10.0,<4.0.0",
    "plotly>=5.24.0,<6.0.0",
    
    # Astronomy and orbital mechanics
    "astropy>=5.3.4,<6.0.0",
    "astroquery>=0.4.11,<0.5.0",
    "poliastro @ git+https://github.com/poliastro/poliastro.git@21fd7719e89a7d22b4eac63141a60a7f1b01768c",
    "sgp4>=2.25,<3.0",
    "jplephem>=2.23,<3.0",
    "pyerfa>=2.0.1.5,<3.0.0",
    
    # ML and embeddings
    "torch>=2.8.0",
    "transformers>=4.57.0,<5.0.0",
    "sentence-transformers>=5.2.0,<6.0.0",
    "huggingface-hub>=0.36.0,<1.0.0",
    "safetensors>=0.7.0,<1.0.0",
    "tokenizers>=0.22.0,<1.0.0",
    "scikit-learn>=1.8.0,<2.0.0",
    "numba>=0.62.0,<1.0.0",
    
    # LangChain ecosystem
    "langchain>=1.1.3,<2.0.0",
    "langchain-core>=1.2.0,<2.0.0",
    "langchain-huggingface>=1.2.0,<2.0.0",
    "langchain-ollama>=1.0.0,<2.0.0",
    "langchain-openai>=1.1.2,<2.0.0",
    "langchain-google-genai>=4.0.0,<5.0.0",
    "langchain-text-splitters>=1.0.0,<2.0.0",
    "langgraph>=1.0.4,<2.0.0",
    "langgraph-checkpoint>=3.0.1,<4.0.0",
    "langgraph-prebuilt>=1.0.5,<2.0.0",
    "langgraph-sdk>=0.2.15,<1.0.0",
    "langsmith>=0.4.59,<1.0.0",
    
    # LLM providers
    "ollama>=0.6.1,<1.0.0",
    "openai>=2.11.0,<3.0.0",
    "google-genai>=1.55.0,<2.0.0",
    
    # Vector database
    "pinecone>=8.0.0,<9.0.0",
    "pinecone-plugin-assistant>=3.0.1,<4.0.0",
    
    # Document processing
    "pypdf>=6.4.1,<7.0.0",
    "beautifulsoup4>=4.14.3,<5.0.0",
    "html5lib>=1.1,<2.0",
    "nltk>=3.9.2,<4.0.0",
    
    # Utilities
    "python-dotenv>=1.2.1,<2.0.0",
    "python-json-logger>=4.0.0,<5.0.0",
    "pydantic>=2.12.5,<3.0.0",
    "tenacity>=9.1.2,<10.0.0",
    "click>=8.3.1,<9.0.0",
    "tqdm>=4.67.1,<5.0.0",
    "requests>=2.32.5,<3.0.0",
    "filelock>=3.20.0,<4.0.0",
    "fsspec>=2025.12.0",
    "packaging>=24.2,<25.0",
    "typing-extensions>=4.15.0,<5.0.0",
    "beartype>=0.22.9,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.24.0,<1.0.0",
    "pytest-cov>=6.0.0,<7.0.0",
    "black>=25.0.0,<26.0.0",
    "ruff>=0.9.0,<1.0.0",
    "mypy>=1.11.0,<2.0.0",
    "pre-commit>=4.5.0,<5.0.0",
    "ipython>=8.30.0,<9.0.0",
    "ipdb>=0.14.0,<1.0.0",
]

gpu = [
    "torch>=2.8.0+cu128",
    "torchvision>=0.23.0+cu128",
]

docs = [
    "mkdocs>=1.6.0,<2.0.0",
    "mkdocs-material>=9.5.0,<10.0.0",
    "mkdocstrings[python]>=0.28.0,<1.0.0",
]

all = [
    "astrobrain[dev,docs]",
]

[project.scripts]
astrobrain = "app.main:main"
orbitqa = "app.orbitqa.orbitqa_app:main"
missionops = "app.missionops.missionops_app:main"

[tool.setuptools]
packages = ["app", "app.core", "app.orbitqa", "app.missionops", "app.rag", "app.astrobrain"]

[tool.setuptools.package-data]
app = ["*.json", "*.yaml", "*.yml"]

# Black configuration
[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | myenv
  | _build
  | buck-out
  | build
  | dist
  | __pycache__
)/
'''

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py312"
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "myenv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "B905",  # zip without explicit strict parameter
    "UP007", # Use X | Y for type annotations (keep Optional for readability)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]
"test_*.py" = ["ARG", "S101"]

[tool.ruff.lint.isort]
known-first-party = ["app"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["app/orbitqa", "app/missionops", "app/rag"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "gpu: marks tests that require GPU",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

# MyPy configuration
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "poliastro.*",
    "astropy.*",
    "sgp4.*",
    "pinecone.*",
    "langchain.*",
    "langgraph.*",
]
ignore_missing_imports = true

# Coverage configuration
[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/myenv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]